changed scripts that run on the action
